Finish frame.c
